+++ R/stat-contour.r
@@ -83,6 +83,7 @@ StatContour <- ggproto("StatContour", Stat,
 
   required_aes = c("x", "y", "z"),
   default_aes = aes(order = after_stat(level)),
+  dropped_aes = "z", # z gets dropped during statistical transformation
 
   setup_params = function(data, params) {
     params$z.range <- range(data$z, na.rm = TRUE, finite = TRUE)
@@ -112,6 +113,7 @@ StatContourFilled <- ggproto("StatContourFilled", Stat,
 
   required_aes = c("x", "y", "z"),
   default_aes = aes(order = after_stat(level), fill = after_stat(level)),
+  dropped_aes = "z", # z gets dropped during statistical transformation
 
   setup_params = function(data, params) {
     params$z.range <- range(data$z, na.rm = TRUE, finite = TRUE)